Renamed the modules and hardcoded the provider code etc
in the controller
we need to change to match the parameters
