+++ app/wizards/publish/a_level_steps/what_a_level_is_required.rb
@@ -1,0 +1,26 @@
+# frozen_string_literal: true
+
+module Publish
+  module ALevelSteps
+    class WhatALevelIsRequired < DfE::Wizard::Step
+      # attr_accessor :answer
+      # validates :answer, presence: true
+
+      # def self.permitted_params
+      #   [:answer]
+      # end
+
+      # def previous_step
+      #   :first_step
+      # end
+
+      # def next_step
+      #   if answer == 'yes'
+      #     :third_step
+      #   elsif answer == 'no'
+      #     :second_step
+      #   end
+      # end
+    end
+  end
+end
